DoubleTriggerPart: migrate from MultipleTriggerPart to just accepting a fixed 2 frames per step
