+++ malcolm/modules/scanning/blocks/double_trigger_block.yaml
@@ -4,7 +4,7 @@
 
 - builtin.parameters.string:
       name: detector
-      description: MRI of the detector to support multiplying triggers
+      description: MRI of the detector to support 2 triggers per step
 
 - builtin.controllers.BasicController:
     mri: $(mri)
