+++ malcolm/modules/scanning/parts/doubletriggerpart.py
@@ -1,6 +1,4 @@
-from typing import Sequence, Union
-
-from annotypes import Anno, Array, add_call_types
+from annotypes import add_call_types
 
 from malcolm.core import CAMEL_RE, APartName, PartRegistrar
 from malcolm.modules import builtin, scanning
@@ -10,20 +8,14 @@ APartName = APartName
 AMri = builtin.parts.AMri
 AInitialVisibility = builtin.parts.AInitialVisibility
 
-with Anno("Single or list of valid multiples"):
-    AMultiples = Union[Array[int]]
 
-
-class MultipleTriggerPart(builtin.parts.ChildPart):
-    """Part for informing a parent that it can have multiple
-    triggers.
-    """
+class DoubleTriggerPart(builtin.parts.ChildPart):
+    """Part for informing a parent that it can have 2 triggers per step."""
 
     def __init__(
         self,
         name: APartName,
         mri: AMri,
-        valid_multiples: Union[AMultiples, Sequence[int], int] = None,
         initial_visibility: AInitialVisibility = False,
     ) -> None:
         super().__init__(
@@ -32,13 +24,6 @@ class MultipleTriggerPart(builtin.parts.ChildPart):
         assert CAMEL_RE.match(name), (
             "MultipleTriggerPart name %r should be camelCase" % name
         )
-        if valid_multiples:
-            if isinstance(valid_multiples, int):
-                self.valid_multiples = [valid_multiples]
-            else:
-                self.valid_multiples = list(valid_multiples)
-        else:
-            self.valid_multiples = list()
 
     def setup(self, registrar: PartRegistrar) -> None:
         super().setup(registrar)
@@ -56,20 +41,6 @@ class MultipleTriggerPart(builtin.parts.ChildPart):
         info = scanning.infos.DetectorMutiframeInfo(detector_mri)
         return info
 
-    def _check_frames_per_step_is_valid(
-        self, detector_mri: str, frames_per_step: int
-    ) -> None:
-        if frames_per_step <= 0:
-            raise ValueError(
-                f"{detector_mri}: frames per step has to be a positive value"
-            )
-        if self.valid_multiples:
-            if frames_per_step not in self.valid_multiples:
-                raise ValueError(
-                    f"{detector_mri}: frames per step {frames_per_step} "
-                    f"not in set of valid values: {self.valid_multiples}"
-                )
-
     # Allow CamelCase as these parameters will be serialized
     # noinspection PyPep8Naming
     @add_call_types
@@ -80,11 +51,13 @@ class MultipleTriggerPart(builtin.parts.ChildPart):
     ) -> None:
         child = context.block_view(self.mri)
         detector_mri = child.detector.value
+        assert (
+            detectors
+        ), f"{detector_mri}: requires a detector table with 2 frames per step"
         if detectors:
             for enable, _, mri, _, frames in detectors.rows():
                 if mri == detector_mri:
-                    if enable:
-                        self._check_frames_per_step_is_valid(detector_mri, frames)
-                    return
-        else:
-            self._check_frames_per_step_is_valid(detector_mri, 1)
+                    if enable and frames != 2:
+                        raise ValueError(
+                            f"{detector_mri}: frames per step has to be equal to 2"
+                        )
